Place successview below navigation bar
